Refactor webgl code such that individual components don't need boilerplate
This pushes a lot of the complexity out of individual components and in to the general "WebGL" comp, as well as creating a nice RenderProgram class to handle a lot of this stuff.  On init, indivdual components now just need to provide the shader/vertex code, and a list of attributes with a bit of metadata.  Then on the "Draw" event, all they need to do is to write the relevent attribute data.